backend/requirements.txt
pyrogram
openai
supabase
python-dotenv
pyyaml
fastapi
uvicorn
